rqt_graph: changed the method of zooming
This new method handles smaller step sizes. The
trackpad on Macbook Pros have a step size which
ranges from +/-2 up to +/-1200, but with a normal
mouse, in OS X or Linux, the step size is always
+/-120 or +/-240. This new method scales better
with smaller step sizes, making scrolling
smoother when using a track pad, which was pretty
much useless before.